LayoutTests/fast/exclusions/parsing/parsing-wrap-flow-expected.txt

Test parsing of the CSS wrap-flow property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("-webkit-wrap-flow", "auto") is "auto"
PASS getCSSText("-webkit-wrap-flow", "both") is "both"
PASS getCSSText("-webkit-wrap-flow", "start") is "start"
PASS getCSSText("-webkit-wrap-flow", "end") is "end"
PASS getCSSText("-webkit-wrap-flow", "maximum") is "maximum"
PASS getCSSText("-webkit-wrap-flow", "clear") is "clear"
PASS getCSSText("-webkit-wrap-flow", ";") is ""
PASS getCSSText("-webkit-wrap-flow", "5") is ""
PASS getCSSText("-webkit-wrap-flow", "-1.2") is ""
PASS getCSSText("-webkit-wrap-flow", "'string'") is ""
PASS getComputedStyleValue("-webkit-wrap-flow", "auto") is "auto"
PASS getComputedStyleValue("-webkit-wrap-flow", "5") is "auto"
PASS getComputedStyleValue("-webkit-wrap-flow", "'string'") is "auto"
PASS getChildComputedStyle("-webkit-wrap-flow", "auto", "start") is "start"
PASS getChildComputedStyle("-webkit-wrap-flow", "end", "auto") is "auto"
PASS getChildComputedStyle("-webkit-wrap-flow", "both", "clear") is "clear"
PASS successfullyParsed is true

TEST COMPLETE
